refactor: created UserWave
