--- packages/client/src/components/HomeMyWave/index.js
@@ -1,1 +1,0 @@
-export { default } from "./HomeMyWave";
