--- packages/client/src/components/HomePopular/index.js
@@ -1,1 +1,0 @@
-export { default } from "./HomePopular";
